Move the factory & AutoRemovingFragment to dedicated package
So that these do not reside in top level library namespace
